make publishing to spec optional

+++ tools/integrations/manage.sh
@@ -7,10 +7,11 @@ set -x
 
 USAGE="
 Usage: $(basename "$0") <cmd>
+For publish, if you want to push the spec to the spec cache, provide a path to a service account key file that can write to the cache.
 Available commands:
   scaffold
   build  <integration_root_path> [<run_tests>]
-  publish  <integration_root_path> [<run_tests>]
+  publish  <integration_root_path> [<run_tests>] [<spec_cache_service_account_key_file_path>]
 "
 
 _check_tag_exists() {
@@ -47,6 +48,7 @@ cmd_publish() {
   [ -d "$path" ] || error "Path must be the root path of the integration"
 
   local run_tests=$1; shift || run_tests=true
+  local spec_cache_service_account_key_file_path=$1; shift || spec_cache_service_account_key_file_path=
 
   cmd_build "$path" "$run_tests"
 
@@ -70,19 +72,25 @@ cmd_publish() {
   docker push "$versioned_image"
   docker push "$latest_image"
 
-  # publish spec to cache. do so, by running get spec locally and then pushing it to gcs.
-  local tmp_spec_file; tmp_spec_file=$(mktemp)
-  docker run --rm "$versioned_image" spec | \
-    # 1. filter out any lines that are not valid json.
-    jq -R "fromjson? | ." | \
-    # 2. grab any json that has a spec in it.
-    # 3. if there are more than one, take the first one.
-    # 4. if there are none, throw an error.
-    jq -s "map(select(.spec != null)) | map(.spec) | first | if . != null then . else error(\"no spec found\") end" \
-    > "$tmp_spec_file"
-  docker run --rm "$versioned_image" spec | jq .spec > "$tmp_spec_file"
-  gcloud auth activate-service-account --key-file /Users/charles/Downloads/prod-ab-cloud-proj-bdb658ebbe5a.json
-  gsutil cp "$tmp_spec_file" gs://io-airbyte-cloud-spec-cache/specs/"$image_name"/"$image_version"/spec.json
+  if [[ -n "${spec_cache_service_account_key_file_path}" ]]; then
+    echo "Publishing and writing to spec cache."
+
+    # publish spec to cache. do so, by running get spec locally and then pushing it to gcs.
+    local tmp_spec_file; tmp_spec_file=$(mktemp)
+    docker run --rm "$versioned_image" spec | \
+      # 1. filter out any lines that are not valid json.
+      jq -R "fromjson? | ." | \
+      # 2. grab any json that has a spec in it.
+      # 3. if there are more than one, take the first one.
+      # 4. if there are none, throw an error.
+      jq -s "map(select(.spec != null)) | map(.spec) | first | if . != null then . else error(\"no spec found\") end" \
+      > "$tmp_spec_file"
+    docker run --rm "$versioned_image" spec | jq .spec > "$tmp_spec_file"
+    gcloud auth activate-service-account --key-file "$spec_cache_service_account_key_file_path"
+    gsutil cp "$tmp_spec_file" gs://io-airbyte-cloud-spec-cache/specs/"$image_name"/"$image_version"/spec.json
+  else
+    echo "Publishing without writing to spec cache."
+  fi
 }
 
 main() {
